move more execution engine code out of workflows package
