+++ packages/core/src/ActiveWorkflows.ts
@@ -22,10 +22,14 @@ import { Service } from 'typedi';
 
 import type { IWorkflowData } from './Interfaces';
 import { ScheduledTaskManager } from './ScheduledTaskManager';
+import { TriggersAndPollers } from './TriggersAndPollers';
 
 @Service()
 export class ActiveWorkflows {
-	constructor(private readonly scheduledTaskManager: ScheduledTaskManager) {}
+	constructor(
+		private readonly scheduledTaskManager: ScheduledTaskManager,
+		private readonly triggersAndPollers: TriggersAndPollers,
+	) {}
 
 	private activeWorkflows: { [workflowId: string]: IWorkflowData } = {};
 
@@ -75,7 +79,8 @@ export class ActiveWorkflows {
 
 		for (const triggerNode of triggerNodes) {
 			try {
-				triggerResponse = await workflow.runTrigger(
+				triggerResponse = await this.triggersAndPollers.runTrigger(
+					workflow,
 					triggerNode,
 					getTriggerFunctions,
 					additionalData,
@@ -150,7 +155,7 @@ export class ActiveWorkflows {
 			});
 
 			try {
-				const pollResponse = await workflow.runPoll(node, pollFunctions);
+				const pollResponse = await this.triggersAndPollers.runPoll(workflow, node, pollFunctions);
 
 				if (pollResponse !== null) {
 					pollFunctions.__emit(pollResponse);
